# specs/api/workspace-sync.openapi.yaml
openapi: 3.1.0
info:
  title: Prodivix Workspace Sync API
  version: 0.2.0
  description: >
    Canonical API for workspace-first editing, partitioned revision recovery,
    and atomic WorkspaceOperation commits. API-003 was amended by the Atomic
    Commit Hard Cut on 2026-07-12; the former /batch contract is removed.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/workspaces/{workspaceId}:
    get:
      summary: Get workspace snapshot
      description: >
        Returns the snapshot only to the workspace owner. A missing workspace
        and a workspace owned by another user both return the same 404 envelope
        so callers cannot enumerate workspace identifiers.
      operationId: getWorkspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWorkspaceResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Workspace not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/workspaces/{workspaceId}/capabilities:
    get:
      summary: Get supported intent and command capabilities
      description: >
        Returns capabilities only to the workspace owner. Authorization
        failures use the same metadata-free 404 envelope as a missing workspace.
      operationId: getWorkspaceCapabilities
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Capability map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCapabilitiesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Workspace not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/workspaces/{workspaceId}/intents:
    post:
      summary: Apply one user intent
      description: >
        Applies an intent only after verifying that the authenticated caller
        owns the workspace. Owner mismatches return a metadata-free 404 and
        never expose current revision details.
      operationId: applyWorkspaceIntent
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyIntentRequest'
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Workspace not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Unsupported or invalid intent envelope, or revision range/capacity failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceConflictErrorEnvelope'
  /api/workspaces/{workspaceId}/operations/commit:
    post:
      summary: Atomically commit one WorkspaceOperation
      description: >
        Commits one command or transaction in a single database transaction.
        The server derives the operation write set, validates the exact
        workspace/route/document revision vector under lock, applies every
        command and reverse/final validator, advances opSeq once, and writes one
        durable operation record. The command or transaction id is the
        idempotency identity: an identical retry returns the first mutation
        result without applying again; reuse with a different request is
        rejected with 422. Workspace settings are outside this endpoint;
        active-document and active-route selections are ephemeral and do not
        form revision preconditions. Selection-only operations and semantic
        no-ops are rejected with 422 because every successful commit must have
        a durable authoring delta. Owner mismatches return a metadata-free 404
        before revision details are built.
      operationId: commitWorkspaceOperation
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceOperationCommitRequest'
      responses:
        '200':
          description: Atomically committed or idempotently replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceOperationCommitSuccessResponse'
        '400':
          description: Malformed JSON, unknown request field, or invalid wire type/format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Workspace not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: >
            Invalid operation, unsupported path, non-exact revision vector,
            revision capacity failure, or operation id reused with a different
            canonical request. Atomic
            commits also reject move/copy, settings, and document-targeted
            asset/project-config mutations, plus operations without a durable
            authoring delta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceConflictErrorEnvelope'
        '500':
          description: Atomic commit could not be completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque session token
  schemas:
    GetWorkspaceResponse:
      type: object
      required: [workspace]
      properties:
        workspace:
          $ref: '#/components/schemas/WorkspaceSnapshot'
      additionalProperties: false
    GetCapabilitiesResponse:
      type: object
      required: [workspaceId, capabilities]
      properties:
        workspaceId:
          type: string
        capabilities:
          type: object
          description: >
            Key format namespace.type@version. Includes both intent and command
            capabilities. Reserved domains (e.g. core.nodegraph.*,
            core.animation.*) may appear as false before corresponding editors
            are implemented.
          additionalProperties:
            type: boolean
    WorkspaceSnapshot:
      type: object
      required:
        [
          id,
          workspaceRev,
          routeRev,
          opSeq,
          tree,
          documents,
          routeManifest,
          settings,
        ]
      properties:
        id:
          type: string
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        opSeq:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        tree:
          $ref: '#/components/schemas/WorkspaceTree'
        documents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/WorkspaceDocument'
        routeManifest:
          $ref: '#/components/schemas/WorkspaceRouteManifest'
        settings:
          type: object
          additionalProperties: true
      additionalProperties: false
    WorkspaceRouteCanonicalString:
      type: string
      minLength: 1
      pattern: '^(?:\S|\S[\s\S]*\S)(?![\s\S])'
      description: Non-empty route identity with no leading or trailing whitespace.
    WorkspaceRouteManifest:
      type: object
      required: [version, root]
      description: >
        Closed canonical RouteManifest wire document. Unknown fields and null
        placeholders are rejected instead of being persisted or discarded.
        Route ids are globally unique, mount ids are unique, and every modules
        map key must exactly equal its value's moduleId. Document references in
        both the project root and every module root are kind-constrained:
        layoutDocId references pir-layout, while pageDocId and outlet binding
        pageDocId reference pir-page or pir-component documents.
      properties:
        version:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        root:
          $ref: '#/components/schemas/WorkspaceRouteRootNode'
        modules:
          type: object
          propertyNames:
            $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          additionalProperties:
            $ref: '#/components/schemas/WorkspaceRouteModule'
        mounts:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceRouteModuleMount'
      additionalProperties: false
    WorkspaceRouteRootNode:
      allOf:
        - $ref: '#/components/schemas/WorkspaceRouteNode'
        - type: object
          properties:
            id:
              type: string
              const: root
...
    WorkspaceTree:
      type: object
      required: [treeRootId, treeById]
      properties:
        treeRootId:
          $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        treeById:
          type: object
          minProperties: 1
          propertyNames:
            $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties:
            $ref: '#/components/schemas/WorkspaceTreeNode'
      additionalProperties: false
    WorkspaceTreeNode:
      oneOf:
        - type: object
          required: [id, kind, name, parentId, children]
          properties:
            id:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            kind:
              type: string
              const: dir
            name:
              type: string
            parentId:
              oneOf:
                - $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
                - type: 'null'
            children:
              type: array
              uniqueItems: true
              items:
                $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties: false
        - type: object
          required: [id, kind, name, parentId, docId]
          properties:
            id:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            kind:
              type: string
              const: doc
            name:
              type: string
            parentId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            docId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties: false
      discriminator:
        propertyName: kind
    WorkspaceDocument:
      type: object
      required: [id, type, path, contentRev, metaRev, content]
      properties:
        id:
          $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        name:
          type: string
          pattern: '\S'
        type:
          $ref: '#/components/schemas/WorkspaceDocumentType'
        path:
          type: string
        contentRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        metaRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        content:
          description: JSON document content. Code documents use the language/source wrapper.
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          description: >
            Omitted when empty. Otherwise values are non-empty canonical strings
            sorted by Unicode code-point order.
          items:
            $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        updatedAt:
          type: string
          format: date-time
          description: >
            Optional in shared authoring-state documents. Server mutation
            acknowledgements require the authoritative timestamp.
      additionalProperties: false
    WorkspaceDocumentType:
      type: string
      enum:
        - pir-page
        - pir-layout
        - pir-component
        - pir-graph
        - pir-animation
        - code
        - asset
        - project-config
...
    ApplyIntentRequest:
      type: object
      required: [expectedWorkspaceRev, intent]
      properties:
        expectedWorkspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        expectedRouteRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        intent:
          $ref: '#/components/schemas/IntentEnvelope'
        clientMutationId:
          type: string
    IntentEnvelope:
      type: object
      required: [id, namespace, type, version, payload, issuedAt]
      properties:
        id:
          type: string
        namespace:
          type: string
          description: core | plugin.<vendor> | custom.<vendor> | core.nodegraph | core.animation
        type:
          type: string
        version:
          type: string
          description: semantic version string, e.g. 1.0
        payload:
          type: object
          additionalProperties: true
        idempotencyKey:
          type: string
        actor:
          type: object
          properties:
            userId:
              type: string
            clientId:
              type: string
          additionalProperties: false
        issuedAt:
          type: string
          format: date-time
...
    WorkspaceOperationCanonicalId:
      type: string
      minLength: 1
      pattern: '^(?:\S|\S[\s\S]*\S)(?![\s\S])'
      description: Non-empty identifier with no leading or trailing whitespace.
...
    WorkspaceOperationCommitRequest:
      type: object
      required: [expected, operation]
      properties:
        expected:
          $ref: '#/components/schemas/WorkspaceOperationCommitExpectedRevisions'
        operation:
          $ref: '#/components/schemas/WorkspaceOperation'
      additionalProperties: false
    WorkspaceOperationCommitExpectedRevisions:
      type: object
      required: [documents]
      dependentRequired:
        routeRev: [workspaceRev]
      description: >
        Exact CAS vector derived from the WorkspaceOperation write set.
        Documents are sorted by id in Unicode code-point order and ids must be
        unique. Omitted partitions are not written by this operation; extra
        partition preconditions are rejected instead of silently
        over-constraining the commit.
      properties:
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        documents:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceOperationCommitExpectedDocument'
      additionalProperties: false
...
    WorkspaceOperation:
      oneOf:
        - $ref: '#/components/schemas/WorkspaceCommandOperation'
        - $ref: '#/components/schemas/WorkspaceTransactionOperation'
      discriminator:
        propertyName: kind
        mapping:
          command: '#/components/schemas/WorkspaceCommandOperation'
          transaction: '#/components/schemas/WorkspaceTransactionOperation'
...
    MutationSuccessResponse:
      type: object
      required: [workspaceId, workspaceRev, routeRev, opSeq]
      properties:
        workspaceId:
          type: string
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        updatedDocuments:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceDocument'
        removedDocumentIds:
          type: array
          items:
            type: string
        tree:
          $ref: '#/components/schemas/WorkspaceTree'
        routeManifest:
          $ref: '#/components/schemas/WorkspaceRouteManifest'
        settings:
          type: object
          additionalProperties: true
        opSeq:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        acceptedMutationId:
          type: string
      additionalProperties: false
    WorkspaceOperationCommitSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/MutationSuccessResponse'
        - type: object
          required: [acceptedMutationId]
          anyOf:
            - required: [tree]
            - required: [routeManifest]
            - required: [updatedDocuments]
            - required: [removedDocumentIds]
          not:
            required: [settings]
          properties:
            workspaceId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            updatedDocuments:
              type: array
              minItems: 1
              items:
                allOf:
                  - $ref: '#/components/schemas/WorkspaceDocument'
                  - type: object
                    required: [updatedAt]
            removedDocumentIds:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            acceptedMutationId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
      description: >
        Aggregated delta for the complete atomic operation. acceptedMutationId
        equals the command or transaction id. An idempotent replay returns the
        originally committed revisions and delta unchanged. Workspace settings
        are not part of an operation commit response. opSeq is monotonic but is
        not a request precondition, so disjoint intervening commits may create a
        sequence gap; this response contains only the current operation delta,
        not a catch-up stream.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    WorkspaceConflictErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          allOf:
            - $ref: '#/components/schemas/ErrorPayload'
            - type: object
              required: [details]
              properties:
                details:
                  $ref: '#/components/schemas/WorkspaceConflictDetails'
      additionalProperties: false
    WorkspaceConflictDetails:
      type: object
      required: [conflictType, workspaceId, expected, current]
      allOf:
        - if:
            properties:
              conflictType:
                const: WORKSPACE_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                allOf:
                  - required: [workspaceRev]
                  - not:
                      required: [document]
              current:
                not:
                  required: [document]
        - if:
            properties:
              conflictType:
                const: ROUTE_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                allOf:
                  - required: [workspaceRev, routeRev]
                  - not:
                      required: [document]
              current:
                not:
                  required: [document]
        - if:
            properties:
              conflictType:
                const: DOCUMENT_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                required: [document]
              current:
                required: [document]
      properties:
        conflictType:
          type: string
          enum: [WORKSPACE_CONFLICT, ROUTE_CONFLICT, DOCUMENT_CONFLICT]
        workspaceId:
          type: string
        expected:
          $ref: '#/components/schemas/WorkspaceConflictExpectedRevisions'
        current:
          $ref: '#/components/schemas/WorkspaceConflictCurrentRevisions'
      additionalProperties: false
...
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - WKS-4001
            - WKS-4002
            - WKS-4003
            - WKS-5001
            - WKS-5002
            - PIR-4001
            - API-1001
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        requestId:
          type: string
        docsUrl:
          type: string
        message:
          type: string
        details:
          description: Safe machine-readable context. Revision conflicts use WorkspaceConflictDetails and never include document content.
        diagnostics:
          type: array
          items:
            $ref: '#/components/schemas/BackendDiagnostic'
    BackendDiagnostic:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        path:
          type: string
        targetRef:
          type: object
          additionalProperties: true
        docsUrl:
          type: string
        details:
          additionalProperties: true
      additionalProperties: true